Harmonisation et nettoyage complet
