+++ src/Form/ResetPasswordType.php
@@ -6,7 +6,6 @@ use App\Entity\User;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -30,11 +29,6 @@ class ResetPasswordType extends AbstractType
                 'second_options' => array('label' => 'Repeat Password'),
                 'required' => true,
                 'mapped' => false,
-            ))
-            ->add('submit', SubmitType::class, array(
-                'attr' => array(
-                    'class' => 'btn btn-primary btn-block'
-                )
             ));
     }
