+++ templates/Front/user/_form.html.twig
@@ -1,5 +1,5 @@
 {{ form_start(form) }}
     {{ form_widget(form) }}
-    <img width="400" src="" alt="" id="profilePicture">
-    <button class="btn btn-primary">{{ button_label|default('Save') }}</button>
+<img width="400" src="" alt="" id="profilePicture">
+<button class="btn btn-primary">{{ button_label|default('Save') }}</button>
 {{ form_end(form) }}
