+++ templates/Front/user/_modal_user_exp_edit.html.twig
@@ -1,4 +1,4 @@
 {{ form_start(formEditExp) }}
     {{ form_widget(formEditExp) }}
-    <button class="btn">{{ button_label|default('Update') }}</button>
+<button class="btn">{{ button_label|default('Update') }}</button>
 {{ form_end(formEditExp) }}
