+++ templates/Front/user/reset_password.html.twig
@@ -7,9 +7,17 @@
 {% endblock %}
 
 {% block content %}
-    <div class="container mt-5 mb-5">
-        <h1>Modifier son mot de passe</h1>
-            {{ form(form) }}
-        <a class="ml-1 mt-3 btn btn-secondary" href="{{ path('front_index') }}">Retour</a>
+    <div class="container-fluid">
+        <div class="row mt-4 mb-4">
+            {{ include('Front/user/_user_info_menu.html.twig') }}
+            <div class="col-lg-9">
+                <h1>Modifier son mot de passe</h1>
+                {{ form_start(form) }}
+                {{ form_widget(form) }}
+                <button class="ml-1 mt-3 btn btn-primary">{{ button_label|default('Save') }}</button>
+                {{ form_end(form) }}
+                <a class="ml-1 mt-3 btn btn-secondary" href="{{ path('front_index') }}">Retour</a>
+            </div>
+        </div>
     </div>
 {% endblock %}
